+++ src/components/ui/modal/ModalButtons.tsx
@@ -1,16 +1,19 @@
 import React from "react";
-import { ModalButton } from "./ModalButton";
+import classnames from "classnames";
 
+import { ModalButton } from "./ModalButton";
 import styles from "./ModalButtons.module.scss";
 
-interface Props {}
+interface Props {
+    className?: string;
+}
 
 interface SubComponents {
     Button: typeof ModalButton;
 }
 
 const ModalButtons: React.FC<Props> & SubComponents = (props) => {
-    return <div className={styles.container}>{props.children}</div>;
+    return <div className={classnames(styles.container, props.className)}>{props.children}</div>;
 };
 ModalButtons.Button = ModalButton;