+++ src/index.ts
@@ -27,3 +27,6 @@ export * from "./components/ui/modal/ModalButtons.js";
 export * from "./components/ui/pop/Pop.js";
 export * from "./components/ui/tabs/Selector.js";
 export * from "./components/ui/toaster/Toaster.js";
+
+export * from "./components/utils/HandleEsc.js";
+export * from "./components/utils/Spacer.js";